Add `Attribute` and `Software`
Remove `MessageTooLarge` error variant

The length of the message is now stored internally in a u16. If the message is too large (i.e. more than u16::MAX bytes) then the program will crash due to an overflow. Realistically this isn't an issue as even if you create a message with every possible attribute, it still won't reach the max size.