Switch to `ring` for cryptography
To generate a certificate from the ed25519 key pair, it must first be converted to an `rcgen::KeyPair` which in turn can only be created from the DER representation of the key. `ed25519-dalek` does not support serializing to the DER format, and so it wouldn't be possible to generate a certificate.

Signed-off-by: Klim Tsoutsman <klimusha@gmail.com>