Use same endpoint for server and peer connections
The server connection and peer connection must use the same socket as
getting a new socket also changes the port number. I preferred the
older API with separate server and client connection types but creating
an endpoint consumes the socket and so we can't just wrap it in an
`Rc<RefCell<T>>`. We could wrap the `Endpoint` and `Incoming` but we
would somehow need to enforce the endpoint having the correct
configuration for its current use (i.e. communicating to server versus
peer).

Signed-off-by: Klim Tsoutsman <klimusha@gmail.com>